# docs/swagger.yaml
components:
  schemas:
    api.GraphSnapshotResponse:
      properties:
        edges:
          items:
            $ref: '#/components/schemas/api.SnapshotEdge'
          type: array
          uniqueItems: false
        metadata:
          $ref: '#/components/schemas/api.SnapshotMetadata'
        nodes:
          items:
            $ref: '#/components/schemas/api.SnapshotNode'
          type: array
          uniqueItems: false
      type: object
...
    api.SnapshotEdge:
      properties:
        id:
          type: string
        latencyP95Ms:
          type: number
        reqRate:
          type: number
        source:
          type: string
        target:
          type: string
      type: object
    api.SnapshotMetadata:
      properties:
        edgeCount:
          type: integer
        edgesWithMetrics:
          type: integer
        generatedAt:
          type: string
        lastUpdatedSecondsAgo:
          type: integer
        nodeCount:
          type: integer
        nodesWithMetrics:
          type: integer
        stale:
          type: boolean
        windowMinutes:
          type: integer
      type: object
    api.SnapshotNode:
      properties:
        availability:
          type: number
        availabilityPct:
          type: number
        betweenness:
          type: number
        errorRatePct:
          type: number
        id:
          type: string
        latencyP95Ms:
          type: number
        name:
          type: string
        namespace:
          type: string
        pageRank:
          type: number
        podCount:
          type: integer
        reqRate:
          type: number
        riskLevel:
          type: string
        riskReason:
          type: string
        updatedAt:
          type: string
      type: object
...
info:
  contact:
    email: support@example.com
    name: API Support
  description: API for the Predictive Analysis Engine, responsible for failure simulation,
    scaling analysis, and risk assessment.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Predictive Analysis Engine API
  version: "1.0"
openapi: 3.1.0
paths:
...
  /dependency-graph/snapshot:
    get:
      description: Fetches the current dependency graph snapshot with optional filtering
        by namespace
      parameters:
      - description: Filter by namespace
        in: query
        name: namespace
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.GraphSnapshotResponse'
          description: OK
        "503":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: Service Unavailable
      summary: Get Dependency Graph Snapshot
      tags:
      - graph
...